fixes required for deeper pymathics modules
