+++ mathics/eval/pymathics.py
@@ -69,10 +69,11 @@ def load_pymathics_module(definitions, module_name: str):
         if var is not None:
             instance = var(expression=False)
             if isinstance(instance, Builtin):
+                submodule_name = var.__module__
                 if not var.context:
                     var.context = "Pymathics`"
                 symbol_name = instance.get_name()
-                builtins_by_module[loaded_module.__name__].append(instance)
+                builtins_by_module.setdefault(submodule_name,[]).append(instance)
                 newsymbols[symbol_name] = instance
 
     for name in newsymbols:
